requirements.txt
beautifulsoup4==4.8.2
bs4==0.0.1
decorator==4.4.1
flake8==3.7.9
Flask==1.1.1
Flask-JWT==0.3.2
Flask-Limiter==1.4
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
gunicorn==20.0.4
HeapDict==1.0.1
html5lib==1.0.1
json5==0.9.1
jsonschema==3.2.0
lxml==4.5.0
MarkupSafe==1.1.1
matplotlib==3.1.3
mysql-connector-python==8.0.21
nltk==3.4.5
numpy==1.18.1
numpydoc==0.9.2
packaging==20.1
pandas==1.0.1
pandocfilters==1.4.2
psutil==5.6.7
py==1.8.1
PyJWT==1.4.2
SQLAlchemy==1.3.13
tables==3.6.1
unicodecsv==0.14.1
update-checker==0.16
urllib3==1.25.8
virtualenv==20.0.21
websocket-client==0.57.0
Werkzeug==1.0.0
widgetsnbextension==3.5.1
WTForms==2.3.1
